chore: rework and improvements on bottomsheet
