+++ lib/app/modules/map/Components/edicao_waypoints.dart
@@ -1,20 +1,21 @@
 import 'dart:io';
 
-import 'package:biketrilhas_modular/app/modules/map/Components/bottom_sheets.dart';
 import 'package:biketrilhas_modular/app/modules/map/map_controller.dart';
 import 'package:biketrilhas_modular/app/modules/usertrails/usertrails_controller.dart';
+import 'package:biketrilhas_modular/app/shared/controller/map_controller.dart';
 import 'package:biketrilhas_modular/app/shared/info/dados_waypoint_model.dart';
 import 'package:biketrilhas_modular/app/shared/info/info_repository.dart';
 import 'package:biketrilhas_modular/app/shared/trilhas/trilha_model.dart';
 import 'package:biketrilhas_modular/app/shared/trilhas/trilha_repository.dart';
 import 'package:biketrilhas_modular/app/shared/utils/functions.dart';
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:photo_view/photo_view.dart';
+
+import 'bottomsheets/bottomsheet_waypoint.dart';
 
 class EdicaoWaypoint extends StatefulWidget {
   final EditMode editMode;
@@ -33,7 +34,7 @@ class _EdicaoWaypointState extends State<EdicaoWaypoint> {
   final _mapController = Modular.get<MapController>();
   final _infoRepository = Modular.get<InfoRepository>();
   final _trilhaRepository = Modular.get<TrilhaRepository>();
-   final _userTrailsController = Modular.get<UsertrailsController>();
+  final _userTrailsController = Modular.get<UsertrailsController>();
 
   @override
   void initState() {
@@ -77,12 +78,11 @@ class _EdicaoWaypointState extends State<EdicaoWaypoint> {
       floatingActionButton: FloatingActionButton(
         child: Icon(Icons.save),
         onPressed: () {
-         
           final m = _mapController.modelWaypoint;
-          if (m.codwp == null){
+          if (m.codwp == null) {
             m.codwp = mapController.nextCodWp();
           }
-          
+
           saida(m);
         },
       ),
@@ -154,7 +154,7 @@ class _EdicaoWaypointState extends State<EdicaoWaypoint> {
           m.codwp, m.codt, m.descricao, m.nome, m.categorias);
       bottomSheetWaypoint(m.codwp);
       Modular.to.pop();
-    } else { 
+    } else {
       //Se o waypoint esta sob gravação
       if (mapController.followTrail != null) {
         mapController.followTrailWaypoints.add(m);
@@ -188,7 +188,7 @@ class _EdicaoWaypointState extends State<EdicaoWaypoint> {
         int codt = await _showTrilhasDialog();
         if (codt != -1) {
           await _showLoadDialog(m, codt);
-          alertEdit(context, "Waypoint salvo com sucesso"); 
+          alertEdit(context, "Waypoint salvo com sucesso");
         }
       }
     }
@@ -258,10 +258,6 @@ class _EdicaoWaypointState extends State<EdicaoWaypoint> {
     );
   }
 
-  
-  
-
-
   Future<int> _showTrilhasDialog() async {
     int _selectedIndex;
     List<TrilhaModel> trilhasproximas;
@@ -316,9 +312,7 @@ class _EdicaoWaypointState extends State<EdicaoWaypoint> {
               FlatButton(
                 child: Text('Cancelar'),
                 onPressed: () {
-                  
                   setState(() {
-                    
                     Navigator.of(context).pop(-1);
                   });
                 },
@@ -577,12 +571,11 @@ class _DialogContentState extends State<DialogContent> {
 }
 
 alertEdit(BuildContext context, String msg) {
-  showDialog(  
+  showDialog(
     context: context,
     barrierDismissible: false,
     builder: (context) {
       return WillPopScope(
-        
         onWillPop: () async => false,
         child: AlertDialog(
           title: Text("Sucesso"),
@@ -592,9 +585,9 @@ alertEdit(BuildContext context, String msg) {
           actions: <Widget>[
             FlatButton(
                 child: Text('OK'),
-                onPressed: () {        
-                  print("Ok, alert edit");    
-                  Navigator.pop(context);                  
+                onPressed: () {
+                  print("Ok, alert edit");
+                  Navigator.pop(context);
                   Modular.to.popUntil((route) => route.isFirst);
                 })
           ],
