+++ lib/app/shared/controller/map_controller.dart
@@ -1,0 +1,4 @@
+import 'package:biketrilhas_modular/app/modules/map/map_controller.dart';
+import 'package:flutter_modular/flutter_modular.dart';
+
+final mapController = Modular.get<MapController>();
